comodines funciona. tener cuidado con la ui

+++ src/main/java/edu/fiuba/algo3/modelo/CondicionesDeMejora/MejoraAlAzar.java
@@ -9,7 +9,7 @@ public class MejoraAlAzar implements IMejora {
     public MejoraAlAzar(int puntos , Random mejora) {}
 
     @Override
-    public void seAplicaApuntaje(PuntajeJugada puntaje, int puntos, int multiplicador) {
+    public void seAplicaApuntaje(PuntajeJugada puntaje, float puntos, float multiplicador) {
 
     }
 }
