+++ src/main/java/edu/fiuba/algo3/modelo/Mejoras/IMejora.java
@@ -3,6 +3,6 @@ import edu.fiuba.algo3.modelo.PuntajeJugada;
 
 public interface IMejora {
 
-	void seAplicaApuntaje(PuntajeJugada puntaje, int puntos, int multiplicador);
+	void seAplicaApuntaje(PuntajeJugada puntaje, float puntos, float multiplicador);
 
 }
