+++ src/main/java/edu/fiuba/algo3/modelo/Mejoras/MultiplicaMultiplicador.java
@@ -9,7 +9,7 @@ public class MultiplicaMultiplicador implements IMejora{
 	}
 
 	@Override
-	public void seAplicaApuntaje(PuntajeJugada puntaje, int puntos, int multiplicador){
+	public void seAplicaApuntaje(PuntajeJugada puntaje, float puntos, float multiplicador){
 		puntaje.multiplicarMultiplicador(multiplicador);
 	}
 }
