+++ src/main/java/edu/fiuba/algo3/modelo/Mejoras/MultiplicaPuntos.java
@@ -8,7 +8,7 @@ public class MultiplicaPuntos implements IMejora{
 	}
 
 	@Override
-	public void seAplicaApuntaje(PuntajeJugada puntaje, int puntos, int multiplicador){
+	public void seAplicaApuntaje(PuntajeJugada puntaje, float puntos, float multiplicador){
 		puntaje.multiplicarPuntos(multiplicador);
 	}
 }
